+++ neo/renderer/BinaryImage.h
@@ -56,6 +56,7 @@ public:
 	void				Load2DAtlasMipchainFromMemory( int width, int height, const byte* pic_const, int numLevels, textureFormat_t& textureFormat, textureColor_t& colorFormat );
 	void				LoadCubeFromMemory( int width, const byte* pics[6], int numLevels, textureFormat_t& textureFormat, bool gammaMips );
 
+	bool				LoadFromGeneratedFile( idFile* f, ID_TIME_T sourceFileTime );
 	ID_TIME_T			LoadFromGeneratedFile( ID_TIME_T sourceFileTime );
 	ID_TIME_T			WriteGeneratedFile( ID_TIME_T sourceFileTime );
 
@@ -123,7 +124,6 @@ private:
 
 private:
 	void				MakeGeneratedFileName( idStr& gfn );
-	bool				LoadFromGeneratedFile( idFile* f, ID_TIME_T sourceFileTime );
 };
 
 #endif // __BINARYIMAGE_H__