+++ neo/swf/SWF_Main.cpp
@@ -29,7 +29,7 @@ If you have questions concerning this license or the applicable additional terms
 #include "precompiled.h"
 #pragma hdrstop
 #include "../renderer/Image.h"
-#include "../renderer/DXT//DXTCodec.h"
+#include "../renderer/DXT/DXTCodec.h"
 
 #pragma warning(disable: 4355) // 'this' : used in base member initializer list